clean incident triage skill and reduce duplication
